==> results/03_pseudotime-monocle3.Rmd <==
---
title: "Meta-analysis of murine lung regeneration studies: monocle3 pseudotime"
author: "Kent Riemondy"
date: "`r Sys.Date()`"
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "html") })
output:
  html_document:
    toc: true
    toc_float: true
    toc_depth: 3
    theme: cosmo
    highlight: kate
    fig_caption: true
    code_folding: hide
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
knitr::opts_chunk$set(cache = FALSE)
```

```{r, results = 'hide', message = FALSE, warning = FALSE}
library(ComplexHeatmap)
library(SingleCellExperiment)
library(scran)
library(scuttle)
library(scater)
library(batchelor)
library(data.table)
library(here)
library(openxlsx)
library(cowplot)
library(monocle3)
library(patchwork)
data_dir <- here("data")
formatted_data_dir <- "processed_data"
fig_dir <- "figs"
obj_dir <- "objects"
mkrs_dir <- "tables"
dir.create("objects/pt", recursive = TRUE, showWarnings = FALSE)
source(here("results/funs.R"))
```

## Pseudotime

Pseudotime analysis will be performed using [`monocle3`](https://cole-trapnell-lab.github.io/monocle3/). The start point of the trajectories will be set to the naive AEC2 cells from the JCI insight manuscript. 

## scvi integration 


```{r, fig.cap="UMAP showing inferred lineages, colored by average pseudotime across all lineages"}
scvi_sce <- readRDS(file.path(obj_dir, "20221230_scvi_sce.rds"))
scvi_sce <- scvi_sce[, scvi_sce$clusters_8 != "8"]
scvi_sce$pt_labels <- ifelse(scvi_sce$cell_type == "Naive AEC2",
                             "Naive AE2",
                             scvi_sce$clusters_8)
if(!file.exists("objects/pt/monocle_scvi.rds")){
  cds <- SeuratObject::as.Seurat(scvi_sce) |> SeuratWrappers::as.cell_data_set()
  
  # note that clustering is done on the UMAP...
  cds <- cluster_cells(cds)
  
  # all cells are in the same lineage for this analysis
  cds@clusters$UMAP$partitions <- setNames(rep(1, length(cds@clusters$UMAP$partitions)),
                                           names(cds@clusters$UMAP$partitions))
  cds <- quiet(learn_graph(cds))
  saveRDS(cds, "objects/pt/monocle_scvi.rds")
} else {
  cds <- readRDS("objects/pt/monocle_scvi.rds")
}
```


```{r, fig.cap = "UMAP showing lineages overlayed on the cell types"}
plot_cells(cds, 
           color_cells_by = "coarse_cell_type",
           label_cell_groups = FALSE,
           label_leaves = TRUE, 
           label_branch_points = FALSE,
           alpha = 0.66) +
  scale_color_manual(values = sccol_pals$tableau10medium)
```


```{r, fig.cap = "UMAP showing lineages overlayed on the clusters"}
plot_cells(cds, 
           color_cells_by = "clusters_8",
           label_cell_groups = FALSE,
           label_leaves = TRUE, 
           label_branch_points = FALSE,
           alpha = 0.66) +
  scale_color_manual(values = sccol_pals$tableau10medium)
```

## scanvi


```{r, fig.cap="UMAP showing inferred lineages, colored by average pseudotime across all lineages"}

scanvi_sce <- readRDS(file.path(obj_dir, "20221230_scanvi_sce.rds"))
scanvi_sce <- scanvi_sce[, scanvi_sce$clusters_9 != "9"]
scanvi_sce$pt_labels <- ifelse(scanvi_sce$cell_type == "Naive AEC2",
                             "Naive AE2",
                             scanvi_sce$clusters_9)

if(!file.exists("objects/pt/monocle_scanvi.rds")){
  cds <- SeuratObject::as.Seurat(scanvi_sce) |> SeuratWrappers::as.cell_data_set()
  
  cds <- cluster_cells(cds)
  cds@clusters$UMAP$partitions <- setNames(rep(1, length(cds@clusters$UMAP$partitions)),
                                           names(cds@clusters$UMAP$partitions))
  cds <- quiet(learn_graph(cds))
  
  saveRDS(cds, "objects/pt/monocle_scanvi.rds")
} else {
  cds <- readRDS("objects/pt/monocle_scanvi.rds")
}
```

```{r, fig.cap = "UMAP showing lineages overlayed on the cell types"}
plot_cells(cds, 
           color_cells_by = "coarse_cell_type",
           label_cell_groups = FALSE,
           label_leaves = TRUE, 
           label_branch_points = FALSE,
           alpha = 0.66) +
  scale_color_manual(values = sccol_pals$tableau10medium)
```

```{r, fig.cap = "UMAP showing lineages overlayed on the clusters"}
plot_cells(cds, 
           color_cells_by = "clusters_9",
           label_cell_groups = FALSE,
           label_leaves = TRUE, 
           label_branch_points = FALSE,
           alpha = 0.66) +
  scale_color_manual(values = sccol_pals$tableau10medium)
```


## harmony


```{r, fig.cap="UMAP showing inferred lineages, colored by average pseudotime across all lineages"}
hm_sce <- readRDS(file.path(obj_dir, "hm_sce.rds"))
hm_sce$pt_labels <- ifelse(hm_sce$cell_type == "Naive AEC2",
                             "Naive AE2",
                             hm_sce$clusters_8)

if(!file.exists("objects/pt/monocle_harmony.rds")){
  cds <- SeuratObject::as.Seurat(hm_sce) |> SeuratWrappers::as.cell_data_set()
  
  cds <- cluster_cells(cds)
  
  cds@clusters$UMAP$partitions <- setNames(rep(1, length(cds@clusters$UMAP$partitions)),
                                           names(cds@clusters$UMAP$partitions))
  cds <- quiet(learn_graph(cds))
  
  saveRDS(cds, "objects/pt/monocle_harmony.rds")
} else {
  cds <- readRDS("objects/pt/monocle_harmony.rds")
}
```

```{r, fig.cap = "UMAP showing lineages overlayed on the cell types"}
plot_cells(cds, 
           color_cells_by = "coarse_cell_type",
           label_cell_groups = FALSE,
           label_leaves = TRUE, 
           label_branch_points = FALSE,
           alpha = 0.66) +
  scale_color_manual(values = sccol_pals$tableau10medium)
```

```{r, fig.cap = "UMAP showing lineages overlayed on the clusters"}
plot_cells(cds, 
           color_cells_by = "pt_labels",
           label_cell_groups = FALSE,
           label_leaves = TRUE, 
           label_branch_points = FALSE,
           alpha = 0.66) +
  scale_color_manual(values = sccol_pals$tableau10medium)
```

## session info

<details><summary>Show session info</summary>

```{r code}
sessionInfo()
```

</details>
